# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Solana War Room API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /wallets/{walletId}/accounts:
    get:
      operationId: listAccountsByWallet
      tags: [accounts]
      summary: List accounts for a wallet
      parameters:
        - name: walletId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Account"
    post:
      operationId: deriveAccounts
      tags: [accounts]
      summary: Derive N accounts from a wallet
      parameters:
        - name: walletId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeriveAccountsRequest"
      responses:
        "201":
          description: Derived accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Account"
...
  /accounts/balances:
    post:
      operationId: refreshBalances
      tags: [accounts]
      summary: Refresh SOL balances for a list of account IDs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshBalancesRequest"
      responses:
        "200":
          description: Updated accounts with balances
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Account"
...
  /arb/configs:
    get:
      operationId: listArbConfigs
      tags: [arb]
      summary: List all arb bot configurations
      responses:
        "200":
          description: List of arb configs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ArbConfig"
    post:
      operationId: createArbConfig
      tags: [arb]
      summary: Create a new arb bot configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateArbConfigRequest"
      responses:
        "201":
          description: Created arb config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArbConfig"
  /arb/configs/{id}:
    delete:
      operationId: deleteArbConfig
      tags: [arb]
      summary: Delete an arb config and its logs
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
  /arb/configs/{id}/start:
    post:
      operationId: startArbConfig
      tags: [arb]
      summary: Start the arb bot for a config
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartArbRequest"
      responses:
        "200":
          description: Updated arb config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArbConfig"
  /arb/configs/{id}/stop:
    post:
      operationId: stopArbConfig
      tags: [arb]
      summary: Stop the arb bot for a config
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated arb config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArbConfig"
  /arb/configs/{id}/logs:
    get:
      operationId: getArbLogs
      tags: [arb]
      summary: Get recent logs for an arb config
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
      responses:
        "200":
          description: List of arb logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ArbLog"
...
components:
  schemas:
...
    Account:
      type: object
      properties:
        id:
          type: integer
        walletId:
          type: integer
        walletName:
          type: string
        name:
          type: string
        publicKey:
          type: string
        encryptedPrivateKey:
          type: string
        hdPath:
          type: string
        hdIndex:
          type: integer
        solBalance:
          type: number
          nullable: true
        selected:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, walletId, walletName, name, publicKey, encryptedPrivateKey, hdPath, hdIndex, selected, createdAt]
    DeriveAccountsRequest:
      type: object
      properties:
        count:
          type: integer
          minimum: 1
          maximum: 100
        password:
          type: string
          description: Password to decrypt mnemonic and re-encrypt private keys
      required: [count, password]
    RefreshBalancesRequest:
      type: object
      properties:
        accountIds:
          type: array
          items:
            type: integer
      required: [accountIds]
...
    ArbConfig:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        accountId:
          type: integer
        accountName:
          type: string
          nullable: true
        publicKey:
          type: string
          nullable: true
        walletName:
          type: string
          nullable: true
        mintAddress:
          type: string
        tokenSymbol:
          type: string
          nullable: true
        inputAmountSol:
          type: number
        minProfitSol:
          type: number
        jitoTipLamports:
          type: integer
        scanIntervalMs:
          type: integer
        slippageBps:
          type: integer
        targetDexes:
          type: array
          items:
            type: string
          description: Jupiter DEX labels to scan (e.g. "Raydium", "Orca", "Whirlpool")
        status:
          type: string
        totalArbs:
          type: integer
        totalProfitSol:
          type: number
        isRunning:
          type: boolean
        startedAt:
          type: string
          nullable: true
        stoppedAt:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, name, accountId, mintAddress, inputAmountSol, minProfitSol, jitoTipLamports, scanIntervalMs, slippageBps, targetDexes, status, totalArbs, totalProfitSol, isRunning, createdAt]
    ArbLog:
      type: object
      properties:
        id:
          type: integer
        configId:
          type: integer
        type:
          type: string
        inputSol:
          type: number
        outputSol:
          type: number
          nullable: true
        profitSol:
          type: number
          nullable: true
        status:
          type: string
        buyDex:
          type: string
          nullable: true
        sellDex:
          type: string
          nullable: true
        buyTxSignature:
          type: string
          nullable: true
        sellTxSignature:
          type: string
          nullable: true
        error:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, configId, type, inputSol, status, createdAt]
    CreateArbConfigRequest:
      type: object
      properties:
        name:
          type: string
        accountId:
          type: integer
        mintAddress:
          type: string
        tokenSymbol:
          type: string
        inputAmountSol:
          type: number
        minProfitSol:
          type: number
          default: 0.001
        jitoTipLamports:
          type: integer
          default: 10000
        scanIntervalMs:
          type: integer
          default: 5000
        slippageBps:
          type: integer
          default: 100
        targetDexes:
          type: array
          items:
            type: string
          description: DEX labels to scan. Defaults to Raydium, Raydium CLMM, Orca, Whirlpool, Meteora DLMM, PumpSwap
      required: [name, accountId, mintAddress, inputAmountSol]
    StartArbRequest:
      type: object
      properties:
        password:
          type: string
      required: [password]
...